+++ src/Shared/DC/ZRDB/tests/test_sqltest.py
@@ -41,15 +41,39 @@ class SQLTestTests(unittest.TestCase):
     def _makeOne(self, *args, **kw):
         return self._getTargetClass()(*args, **kw)
 
-    def test_simple(self):
+    def test_string(self):
         tested = self._makeOne('foo type="string"')
         self.assertEqual(tested.render(FauxMultiDict(foo='FOO')),
                          'foo = "FOO"')
 
-    def test_binary(self):
+    def test_string_binary(self):
         tested = self._makeOne('foo type="string"')
         self.assertEqual(tested.render(FauxMultiDict(foo=b'FOO')),
                          'foo = "FOO"')
+
+    def test_int(self):
+        tested = self._makeOne('foo type="int"')
+
+        for valid in (100, 100.00, '100', b'100'):
+            self.assertEqual(tested.render(FauxMultiDict(foo=valid)),
+                             'foo = 100')
+
+        for invalid in ('', b'', None):
+            self.assertRaises(ValueError,
+                              tested.render,
+                              FauxMultiDict(foo=invalid))
+
+    def test_float(self):
+        tested = self._makeOne('foo type="float"')
+
+        for valid in (100, 100.0, '100.0', b'100.0'):
+            self.assertEqual(tested.render(FauxMultiDict(foo=valid)),
+                             'foo = 100.0')
+
+        for invalid in ('', b'', None):
+            self.assertRaises(ValueError,
+                              tested.render,
+                              FauxMultiDict(foo=invalid))
 
 
 def test_suite():